% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.fitmodel.r
\name{anova.fitmodel}
\alias{anova.fitmodel}
\title{Analysis of Variance of fitted model}
\usage{
\method{anova}{fitmodel}(object, ...)
}
\arguments{
\item{object}{fitmodel object}

\item{...}{other parameters passed to \code{anova}}
}
\description{
Produces an analysis of variance of the "fitmodel" object
}
