% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burd_rain.r
\docType{data}
\name{burd_rain}
\alias{burd_rain}
\title{burd_rain}
\format{An object of class \code{data.frame} with 2492 rows and 2 columns.}
\usage{
burd_rain
}
\description{
Burdekin River rainfall data
}
\details{
This dataset consists of rainfall measurements from a gauge near
the Burdekin. This is a 2492 x 2 data frame containing rainfall
measurements in mm for each date spanning the date ranges
31st December 2010 and 26th of October 2017.
}
\keyword{datasets}
