% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burdRC.r
\docType{data}
\name{burdRC}
\alias{burdRC}
\title{burdRC}
\format{
Burdekin total suspended solids (TSS) data:
Dataframe consisting of 461 rows and 2 columns
\describe{
  \item{Date}{Date, where the range of dates span 2016-01-26 to 2016-06-23}
  \item{TSS}{Daily total suspended sediment (TSS) in mg/L}
  }
}
\usage{
burdRC
}
\description{
Burdekin River concentration (TSS) data
}
\keyword{datasets}
