% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burdRNA.r
\docType{data}
\name{burdRNA}
\alias{burdRNA}
\title{burdRNA}
\format{
An object of class \code{data} of length 2.
}
\usage{
burdRNA
}
\description{
Burdekin River data with missing values
}
\details{
This dataset is a replicate of the burdR R object but
with records missing that span the following date periods
2004-01-13 to 2004-04-22.
}
\keyword{datasets}
