% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burdRQ.r
\docType{data}
\name{burdRQ}
\alias{burdRQ}
\title{burdRQ}
\format{
Burdekin flow data:
Dataframe consisting of 15,186 rows and 2 columns
\describe{
  \item{Date}{Date, where the range of dates span 1973-12-02 to 2015-06-30}
  \item{Flow}{Daily flow in cubic metres per second.}
  }
}
\usage{
burdRQ
}
\description{
Burdekin River Flow data
}
\keyword{datasets}
