% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcSMK.r
\name{CalcSMK}
\alias{CalcSMK}
\title{Calculating SMK}
\usage{
CalcSMK(loaddata, alternative = "two.sided")
}
\arguments{
\item{loaddata}{list from LRE that contains the data}

\item{alternative}{the alternative hypothesis.  This can be changed to test for}
}
\description{
Calculating SMK using regularised data
}
