% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCQDataset.r
\name{CreateCQDataset}
\alias{CreateCQDataset}
\title{CreateCQDAtaset}
\usage{
CreateCQDataset(Conc, Qreg, startdate)
}
\arguments{
\item{Conc}{Concentration data}

\item{Qreg}{Regularised flow data}

\item{startdate}{Starting date}
}
\description{
LRE Internal function
}
