% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateData.R
\name{CreateData}
\alias{CreateData}
\title{Creates datasets that will be used in loads modelling
  and prediction}
\usage{
CreateData(
  Q,
  Conc,
  date.range = list(model = NULL, pred = NULL, hour = FALSE),
  samp.unit = "hour",
  Ytype = "WY",
  Qflush = 0.9,
  Reg = list(type = "ss", rainfall = NULL, date = NULL)
)
}
\arguments{
\item{Q}{\code{n x 2} matrix of flow (Date, Flow) where flow is measured in cumecs}

\item{Conc}{$nxp$ matrix of concentrations  (Date, Conc1, Conc2, ...) where
concentration is measured in mg/L}

\item{date.range}{list containing date ranges for modelling and prediction
datasets (default: NULL)}

\item{samp.unit}{sampling unit 'hour' or 'daily'}

\item{Ytype}{water year "WY" or financial year "FY" (note only WY
implemented at this stage)}

\item{Qflush}{Qflush}

\item{Reg}{Regularisation check}
}
\value{
a \code{list} of class 'regdata' containing:
\code{CQ} flow and concentration modelling dataset
\code{Qreg} regularised dataset
\code{Q} Observed flow matrix
Standard \code{summary} and \code{plot} functions apply.
}
\description{
Creates a regularised flow dataset for prediction and a
modelling dataset with flow metrics that can be used in regression
modelling.
}
\details{
CreateData

This function creates two datasets that are used in
modelling and prediction. A third dataset is just a copy of
the observed flow data.

Inputs into the function are flow and concentration datasets that
have been read in using the \code{ReadInData} function or
are of class "data".
}
\examples{
#Reading in data already stored in R
library(LRE)
burdR <- ReadInDataFromR(x.C = burdRC, x.Q = burdRQ)

# Create Dataset (no need for regularisation here as have full data)
date.rangeM <- c("1973-12-02", "2015-06-30")
date.rangeP <- c("1973-12-02", "2015-06-30")
loaddata <- CreateData(Q = burdR$Q, Conc = burdR$Conc,
                date.range = list(model = date.rangeM, pred = date.rangeP, hour = FALSE),
                samp.unit = "day", Ytype = "WY", Qflush = 0.9,
                Reg = list(type = "none", rainfall = NULL, date = NULL))
p <- plot(loaddata, Type = "WY")
summary(loaddata)

}
\references{
Kuhnert, P.M., Henderson, B.L., Lewis, S.E.,
Bainbridge, Z.T., Wilkinson, S.N. and Brodie, J.E. (2012)
Quantifying total suspended sediment export from the Burdekin
River catchment using the loads regression estimator tool,
Water Resources Research, 48, W04533,doi:10.1029/2011WR011080.
}
\seealso{
ReadInData
}
\keyword{loads,}
\keyword{metrics}
