% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateLD.R
\name{CreateLD}
\alias{CreateLD}
\title{CreateLD}
\usage{
CreateLD(
  Q,
  Conc,
  date.range,
  samp.unit,
  Ytype,
  Qflush,
  Reg = list(type = "none", rainfall = NULL, date = NULL)
)
}
\arguments{
\item{Q}{Flow data}

\item{Conc}{Concentration data}

\item{date.range}{Date range}

\item{samp.unit}{Sampling unit}

\item{Ytype}{Year type. One of water year ("WY" or "WY2") or financial year ("FY")}

\item{Qflush}{Quantile for "first flush"}

\item{Reg}{Parameters for regularisation (type, rainfall, date)}
}
\description{
LRE Internal function
}
