% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateLoadVars.R
\name{CreateLoadVars}
\alias{CreateLoadVars}
\title{CreateLoadVars}
\usage{
CreateLoadVars(Q, csQ, Qflush, samp.unit, Date, Y)
}
\arguments{
\item{Q}{Flow}

\item{csQ}{Cumulative sum of flow}

\item{Qflush}{quantile associated with first flush}

\item{samp.unit}{sampling unit}

\item{Date}{Date vector}

\item{Y}{Year type}
}
\description{
LRE Internal function which create load features
}
