% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateQregDataset.r
\name{CreateQregDataset}
\alias{CreateQregDataset}
\title{CreateQregDataset}
\usage{
CreateQregDataset(Q, samp.unit, Qflush, Ytype, Reg)
}
\arguments{
\item{Q}{Flow}

\item{samp.unit}{sampling unit}

\item{Qflush}{quantile corresponding to first flush}

\item{Ytype}{Year type: water year ("WY"), financial year ("FY")}

\item{Reg}{Regularisation parameters}
}
\description{
LRE Internal function.
}
