% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRegDate.R
\name{CreateRegDate}
\alias{CreateRegDate}
\title{CreateRegDate}
\usage{
CreateRegDate(Q, samp.unit, Ytype)
}
\arguments{
\item{Q}{Flow}

\item{samp.unit}{sampling unit}

\item{Ytype}{Year type: water year ("WY"), financial year ("FY")}
}
\description{
LRE Internal function.
}
