% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateY.R
\name{CreateY}
\alias{CreateY}
\title{Creates a categorical year variable}
\usage{
CreateY(Date, Ytype)
}
\arguments{
\item{Date}{date vector}

\item{Ytype}{year type, either 'WY' (water year) or 'FY'
financial year}
}
\description{
Creates a categorical variable of year
type that is either based on a water year or financial
year.
}
