% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R, R/diagnostic.fitmodel.R
\name{diagnostic}
\alias{diagnostic}
\alias{diagnostic.fitmodel}
\title{Produces diagnostic plots of a fitted loads model}
\usage{
diagnostic(object)

\method{diagnostic}{fitmodel}(object)
}
\arguments{
\item{object}{object of class "fitmodel" resulting from
the \code{\link{FitModel}} function.}
}
\value{
Produces the following diagnostics plots:
\code{residuals versus predicted} points should appear randomly
   distributed along the line at zero.
\code{normal Quantile-Quantile (Q-Q) plot} compares the quantiles
   from the theoretical distribution (i.e. normal) to that observed.
   If the 2 distributions are similar, then the points in the Q-Q plot
   will approximately lie on the line. Som deviation at the tails is
   acceptable.
\code{histogram of residuals} should appear roughly symmetric and bell
  shaped. No skewness (elongation of the tails) or kurtosis (thickening
  of the tails) should be present.
\code{predicted versus observed} points should appear close to the
  diagonal 0-1 line.
\code{predict versus obsered (WY)} plots of predicted versus observed
  for each water year.
\code{residuals versus predicted (WY)} plots of of residuals versus
  predicted for each water year.

 @importFrom "stats" "resid"
 @importFrom "stats" "lag"

 @references Kuhnert, P.M., Henderson, B.L., Lewis, S.E.,
 Bainbridge, Z.T., Wilkinson, S.N. and Brodie, J.E. (2012) Quantifying
 total suspended sediment export from the Burdekin River catchment using
 the loads regression estimator tool, Water Resources Research, 48,
 W04533,doi:10.1029/2011WR011080.

 @seealso FitModel, qqnorm,, qqline

 @keywords loads, diagnostics
}
\description{
S3 method for evaluating diagnostics of a fitted LRE object

Produces a series of standard suite of diagnostics
for examining the fit of a loads model.
}
\details{
Series of diagnostic plots for examining model fit.
}
