% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_trend.r
\name{extract_trend}
\alias{extract_trend}
\title{Extracts the trend}
\usage{
extract_trend(lin_object, lin_term = "trendY", percent = FALSE)
}
\arguments{
\item{lin_object}{gam model produced from the mgcv R package with a linear trend fitted}

\item{lin_term}{name of the linear term}

\item{percent}{Should this be calculated as a percentage (TRUE/FALSE)}
}
\description{
Extracts the trend and 95% confidence intervals from the linear model
}
