% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitModel.R
\name{FitModel}
\alias{FitModel}
\title{FitModel}
\usage{
FitModel(x, parms = list(flow = FALSE, seasonal = FALSE,
          RFlimb = FALSE, MA = c(MA1day = FALSE, MA2days = FALSE,
          MAweek = FALSE, MAmonth = FALSE, MA6months = FALSE,
          MA12months = FALSE), trend = FALSE, correlation = FALSE))
}
\arguments{
\item{x}{The loads dataset, CQ object resulting from the \code{\link{CreateData}}
function.}

\item{parms}{A list comprising terms to fit in a model:
flow: linear (\code{"linear"}) and quadratic (\code{"quadratic"}) terms for flow.
Default: \code{FALSE}
seasonal: logical term indicating whether seasonal terms are fitted. Default:\code{FALSE}
RFlimb: logical term indicating whether a rising/falling limb term is fitted.
Default: \code{FALSE},
MA: vector indicating whether one or more moving average terms are fitted.
Default:\code{c(MA1day = FALSE, MA2days = FALSE, MAweek = FALSE, MAmonth = FALSE,
MA6months = FALSE, MA12months = FALSE)}
trend: logical term indicating whether a long term trend is fitted. Default: \code{FALSE}
correlation: AR1 correlation structure incorporated. Default: \code{FALSE}}
}
\value{
an object of class \code{fitmodel} with elements as described in
\code{gamObject} from the \code{mgcv} package. Standard \code{summary}
and \code{anova} functions apply.
}
\description{
This function fits a generalized additive model (\code{gam})
to a loads dataset.  If correlation structures need to be fitted, especially
to large datasets, do not use this wrapper and opt for a manual fit using \code{bam}.
}
\details{
This is a wrapper function for fitting a generalised additive model using the
\code{link{mgcv}} package. Terms are fitted in the model by "switching" terms on and off.
Apart from the \code{flow} term, all other terms are logical. If a term is intended to be
included in a model, then that term can be set to \code{TRUE}.

To include additional terms to those specified, the user will need to run the \code{gam}
function from the \code{\link{mgcv}} package that this method is based upon. The following is
an example where an additional variable, "dam" is included into the model.
\code{mod <- gam(log(Conc) ~ log(pQ) + I(log(pQ)^2) + per(days, k=2) +  s(MA1day) + s(trend) + dam,
      family = gaussian, data = loaddata$CQ, control = gam.control(keepData = TRUE)) }
      where \code{loaddata$CQ} contains an additional column of data labelled, "dam".
}
\references{
Kuhnert, P.M., Henderson, B.L., Lewis, S.E.,
Bainbridge, Z.T., Wilkinson, S.N. and Brodie, J.E. (2012)
Quantifying total suspended sediment export from the Burdekin
River catchment using the loads regression estimator tool,
Water Resources Research, 48, W04533,doi:10.1029/2011WR011080.
}
\seealso{
ReadInData, CreateData, gam, diagnostic
}
\keyword{gam,}
\keyword{loads}
