% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitSmQ.R
\name{FitSmQ}
\alias{FitSmQ}
\title{FitSmQ}
\usage{
FitSmQ(Q, Qreg, method = "monoH.FC")
}
\arguments{
\item{Q}{flow}

\item{Qreg}{regularised flow}

\item{method}{spline method for interpolation}
}
\description{
LRE Internal function.
}
