% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LQflow.r
\name{LQflow}
\alias{LQflow}
\title{Create linear and quadratic terms for flow}
\usage{
LQflow(pQ, quad = TRUE)
}
\arguments{
\item{pQ}{flow}

\item{quad}{should quadratic terms be created? (Default: TRUE)}
}
\description{
Function used in 'FitModel' to create linear
and quadratic terms for flow.
}
