% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRE-CreateData-Internal.r
\name{LRE-Create-Internal}
\alias{LRE-Create-Internal}
\alias{CreateLD}
\alias{CreateQregDataset}
\alias{PlotCQ}
\alias{CreateCQDataset}
\alias{AssignVal}
\alias{CreateRegDate}
\alias{CreateLoadVars}
\title{Internal functions for \code{CreateData} function.}
\usage{
CreateLD(
  Q,
  Conc,
  date.range,
  samp.unit,
  Ytype,
  Qflush,
  Reg = list(type = "none", rainfall = NULL, date = NULL)
)

CreateQregDataset(Q, samp.unit, Qflush, Ytype, Reg)

PlotCQ(Conc, Q)

CreateCQDataset(Conc, Qreg, startdate)

AssignVal(x, y)

CreateRegDate(Q, samp.unit, Ytype)

CreateLoadVars(Q, csQ, Qflush, samp.unit, Date, Y)
}
\arguments{
\item{Q}{Flow data}

\item{Conc}{Concentration data}

\item{date.range}{Date range}

\item{samp.unit}{sampling unit}

\item{Ytype}{Year type. One of water year ("WY" or "WY2") or financial year ("FY")}

\item{Qflush}{Quantile for "first flush"}

\item{Reg}{Parameters for regularisation (type, rainfall, date)}

\item{Qreg}{Regularised flow data}

\item{startdate}{Starting date}

\item{x}{vector of x-values}

\item{y}{vector of y-values}

\item{csQ}{Cumulative sum of flow}

\item{Date}{Date vector}

\item{Y}{Year type}
}
\description{
Internal functions for curating the data
}
