% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRE-DerivFns-Internal.r
\name{LRE-DerivFns-Internal}
\alias{LRE-DerivFns-Internal}
\alias{Deriv}
\alias{confint.Deriv}
\alias{signifD}
\title{Internal functions for \code{periods_of_change} function.}
\arguments{
\item{object}{gam model object}

\item{n}{number of iterations to calculate derivatives. Default is 200.}

\item{eps}{A small positive constant representing the convergence threshold.}

\item{data}{data used in the gam model.}

\item{object.vars}{variables used in the gam fit.}

\item{plot.Deriv}{plot.Deriv function}

\item{term}{varible in the model object that is of interest.  Typically trend term.}

\item{alpha}{significance level. Default is 0.05.}

\item{...}{additional arguments to base plot function}

\item{x}{predicted value from model}

\item{d}{calculated derivative}

\item{upper}{upper confidence interval}

\item{lower}{lower confidence interval}

\item{eval}{evaluation point.  Defaults to 0.}
}
\description{
Internal functions for calculating and plotting trends on a 
             non-linear curve. Reference for this concept appears in the link below by Gavin Simpson:
}
\keyword{internal}
