% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRE-FitModel-Internal.r
\name{LRE-FitModel-Internal}
\alias{LRE-FitModel-Internal}
\title{Internal functions for \code{FitModel}}
\arguments{
\item{pQ}{Flow record}

\item{quad}{Do you want a quadratic term? (Default: TRUE)}
}
\description{
Internal functions used in FitModel
}
\details{
Internal functions used to fit the model
}
