% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRE-plotpredict-Internal.r, R/LoadPlot.R
\name{LRE-plotpredict-Internal}
\alias{LRE-plotpredict-Internal}
\title{Internal function used to plot predictions in the
\code{plot.PredictLoad} function.}
\arguments{
\item{x}{x}

\item{Conc}{Concentration}

\item{scale}{scale}
}
\description{
Function that calculates confidence intervals

Function that calculates confidence intervals
}
