% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRE-predict-Internal.r
\name{LRE-predict-Internal}
\alias{LRE-predict-Internal}
\alias{Intervals}
\alias{calcSE}
\alias{BlockSigma}
\alias{ConvertFacts}
\alias{EstChat}
\alias{EstLoadYS}
\alias{ExtractDailyLoads}
\alias{LoadPlot}
\title{Internal functions for \code{predict.fitmodel} function.}
\arguments{
\item{est}{Estimation of response from model}

\item{Xdesign}{Design matrix from model}

\item{LhatC}{Estimated concentrations}

\item{alpha1}{error in flows}

\item{alpha2}{error in flows}

\item{beta1}{beta1}

\item{rho}{correlation term}

\item{W}{W matrix}

\item{sig2}{error}

\item{samp.unit}{sampling unit}

\item{x}{x}

\item{Conc}{Concentration}

\item{scale}{scale}

\item{pvalue}{pvalue}

\item{len}{len}

\item{Qdat}{Qdat}

\item{modobj}{modobj}

\item{modobjfix}{modobjfix}

\item{CQdat}{CQdat}

\item{se}{se}

\item{Blocklen}{Blocklen}

\item{concY}{concY}

\item{Date}{Date}

\item{Lhat}{Lhat}

\item{Qhat}{Qhat}

\item{error}{error}

\item{modparms}{modparms}
}
\description{
Internal functions used for prediction.
EstChat: estimates concentration.
BlockSigma: estimates the variance-covariance matrix.
CalcSE: calls on C code to calculate the standard error.
ConvertFacts: Internal function that determines the conversion
constant to use for loads and flow.Conversion to Million tonnes:
C = mg/L  Q = m^3/s (K.loads) mg/l x m3/1000 = 1kg  & 1kg = 0.001 t
Conversion to millions of ML (K.flow).
EstLoadYS: Function used in \code{predict.fitmodel} to estimate the
load based on predicted concentrations from predicted hourly flow records.
EstLoad: Function called by \code{EstLoadYS}
ExtractDailyLoads: extracts the daily loads.
Intervals: calculates confidence intervals.
}
\keyword{internal}
