% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRE-ReadData-Internal.r
\name{LRE-Read-Internal}
\alias{LRE-Read-Internal}
\alias{ReadQ}
\alias{ReadC}
\alias{ExtractDT}
\title{Internal functions for \code{ReadInData} function.}
\arguments{
\item{fn}{file name}

\item{format}{format for the date}

\item{Cnames}{concentration column label}

\item{OmitCzeros}{a logical value indicating whether zeros should
be omitted in the concentration record. If missing, it is assumed
that no zeros should be omitted from the record if they exist.}

\item{dt}{character representing the date}
}
\description{
Internal functions for reading in data
}
\keyword{internal}
