% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.r
\name{LRE-regularize-Internal}
\alias{LRE-regularize-Internal}
\title{features}
\arguments{
\item{series}{series}

\item{dateTime}{dateTime}
}
\description{
Creates rainfall feature variables for prediction of flow in a
random Forest
}
