% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRE-package.r
\docType{package}
\name{LRE}
\alias{LRE-package}
\alias{LRE}
\title{Loads Regression Estimator}
\description{
Estimates loads using a generalised rating curve approach with uncertainties.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkuhnert.github.io/LRE/}
  \item \url{https://github.com/pkuhnert/LRE}
}

}
\keyword{internal}
