% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periods_of_change.r
\name{periods_of_change}
\alias{periods_of_change}
\title{Calculate and Plot Periods of Change in a non-linear curve}
\usage{
periods_of_change(
  fit_sm,
  fit_lin = NULL,
  Term = "trendY",
  data,
  object.vars = c("pQ", "MA2days", "MAweek", "MAmonth", "MA6months", "MA12months",
    "trendY"),
  spacing = 5
)
}
\arguments{
\item{fit_sm}{gam object with smoother for trend produced from the mgcv R package}

\item{fit_lin}{gam object with linear trend term produced from the mgcv R package (Default: NULL)}

\item{Term}{term to calculate the periods of change over, typically a trend term}

\item{data}{data used in the gam model, eg loaddata$CQ}

\item{object.vars}{variables used in the gam fit.}

\item{spacing}{white space left between the labels. Defaults to 5.}
}
\description{
Calculates and plots the episodic trends in a non-linear trend
       curve produced from a gam
}
