% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.data.R
\name{plot.data}
\alias{plot.data}
\title{Plot data objects}
\usage{
\method{plot}{data}(x, logt = FALSE, ...)
}
\arguments{
\item{x}{data object}

\item{logt}{log transformation. Default: FALSE}

\item{...}{other parameters passed to \code{plot}}
}
\description{
Plot object of class 'data'
}
