% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fitmodel.R
\name{plot.fitmodel}
\alias{plot.fitmodel}
\title{Plot predictions from the fitted model}
\usage{
\method{plot}{fitmodel}(x, Qreg, data, ...)
}
\arguments{
\item{x}{data object}

\item{Qreg}{Regularised flow record}

\item{data}{dataset that was used for modelling}

\item{...}{other parameters passed to \code{plot}}
}
\description{
Plot object of class 'fitmodel'
}
