% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PredictLoad.R
\name{plot.PredictLoad}
\alias{plot.PredictLoad}
\title{plot.PredictLoad}
\usage{
\method{plot}{PredictLoad}(x, type = "annual", Conc, newQ = NULL, scale = "Mt", ...)
}
\arguments{
\item{x}{object}

\item{type}{type of load (annual or daily)}

\item{Conc}{Concentration label}

\item{newQ}{new flow vector (Default: NULL)}

\item{scale}{scale to plot loads (Default: Mt)}

\item{...}{other parameters passed to \code{plot}}
}
\description{
Plots load estimates from the fitted model
}
\references{
Kuhnert, P.M., Henderson, B.L., Lewis, S.E.,
Bainbridge, Z.T., Wilkinson, S.N. and Brodie, J.E. (2012)
Quantifying total suspended sediment export from the Burdekin
River catchment using the loads regression estimator tool,
Water Resources Research, 48, W04533,doi:10.1029/2011WR011080.
}
\keyword{loads,}
\keyword{metrics}
