% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.regdata.R
\name{plot.regdata}
\alias{plot.regdata}
\title{Plot regularised data}
\usage{
\method{plot}{regdata}(x, Type, Qcutoff = NULL, ...)
}
\arguments{
\item{x}{x}

\item{Type}{Year type: water year (WY) or financial year (FY)}

\item{Qcutoff}{Qcutoff}

\item{...}{other parameters passed to \code{plot}}
}
\description{
plots an object of class "regdata"
}
