% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictL.fitmodel.r
\name{predictL.fitmodel}
\alias{predictL.fitmodel}
\title{predictL.fitmodel}
\usage{
\method{predictL}{fitmodel}(object, objfix = NULL, x,
  flow.error = list(me = 0, ce = 0), samp.unit = "hour",
  pvalue = 0.05)
}
\arguments{
\item{object}{GAM modelling object resulting from
\code{\link{FitModel}} function.}

\item{objfix}{GAM modelling object resulting from \code{\link{FitModel}}
function when a correlation structure has been implemented.}

\item{x}{load data resulting from the \code{\link{CreateData}}
function.}

\item{flow.error}{a list containing the flow measurement and
cross-sectional error in terms of the coefficient of variation.
Default: none.}

\item{samp.unit}{Sampling unit for the loads calculation, "day" or "hour"}

\item{pvalue}{for confidence intervals. Default: 0.05 (95 %
confidence intervals)}
}
\value{
a \code{list} of class 'PredictLoad' with elements
\code{annual} annual loads estimates
\code{daily} daily loads estimates
 Standard \code{print} and \code{plot} functions apply.
}
\description{
Predicts the load and provides
uncertainty estimates for each water year.
}
\details{
Computes the loads estimates for each water year with standard
error estimates according to the p-value, sampling unit and flow
error specified. By default, the error due to flow rate is assumed
to be zero. Error can be included as a coefficient of variation
according to whether the error is due to the measurement of flow
(me) or the position of the gauge in the river, which is referred
to as "cross-sectional" (ce) error.
}
\note{
Only water year calculations have been implemented.
}
\references{
Kuhnert, P.M., Henderson, B.L., Lewis, S.E.,
Bainbridge, Z.T., Wilkinson, S.N. and Brodie, J.E. (2012)
Quantifying total suspended sediment export from the Burdekin
River catchment using the loads regression estimator tool,
Water Resources Research, 48, W04533,doi:10.1029/2011WR011080.
}
\seealso{
ReadInData, CreateData, FitModel
}
\keyword{uncertainty}
