% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictL.r
\name{predictL}
\alias{predictL}
\title{predictL}
\usage{
predictL(
  object,
  objfix = NULL,
  x,
  flow.error = list(me = 0, ce = 0),
  samp.unit = "hour",
  pvalue = 0.05
)
}
\arguments{
\item{object}{GAM modelling object resulting from
\code{\link{FitModel}} function.}

\item{objfix}{GAM modelling object resulting from \code{\link{FitModel}}
function when a correlation structure has been implemented.}

\item{x}{load data resulting from the \code{\link{CreateData}}
function.}

\item{flow.error}{a list containing the flow measurement and
cross-sectional error in terms of the coefficient of variation.
Default: none.}

\item{samp.unit}{Sampling unit for the loads calculation, "day" or "hour"}

\item{pvalue}{for confidence intervals. Default: 0.05 (95 %
confidence intervals)}
}
\description{
Predicts the Load
}
