% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PredictLoad.R
\name{print.PredictLoad}
\alias{print.PredictLoad}
\title{print.PredictLoad}
\usage{
\method{print}{PredictLoad}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{other arguments passed to \code{print}}
}
\description{
Outputs estimates to file
}
\references{
Kuhnert, P.M., Henderson, B.L., Lewis, S.E.,
Bainbridge, Z.T., Wilkinson, S.N. and Brodie, J.E. (2012)
Quantifying total suspended sediment export from the Burdekin
River catchment using the loads regression estimator tool,
Water Resources Research, 48, W04533,doi:10.1029/2011WR011080.
}
