% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sumregdata.r
\name{print.sumregdata}
\alias{print.sumregdata}
\title{Print out summaries from regularised data}
\usage{
\method{print}{sumregdata}(x, ...)
}
\arguments{
\item{x}{regdata object}

\item{...}{other parameters passed to \code{summary}}
}
\description{
Print summarises an object of class "regdata"
}
