% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizeFlow.r
\name{regularization}
\alias{regularization}
\alias{regularizeFlowSS}
\alias{regularizeFlowR}
\title{Internal functions to create a regularized flow
dataset using either a smoothing spline or rainfall record}
\usage{
regularizeFlowSS(Q, Qreg)

regularizeFlowR(rainfall, dateTimeRainfall, flow, dateTimeFlow)
}
\arguments{
\item{Q}{Q}

\item{Qreg}{Qreg}

\item{rainfall}{rainfall}

\item{dateTimeRainfall}{dateTimeRainfall}

\item{flow}{flow}

\item{dateTimeFlow}{dateTimeFlow}
}
\description{
Creates a regularised flow dataset
using a smoothing spline or rainfall record using
Random Forests. Predictions produced from the smoothing
spline approach are checked to ensure predictions do not
go outside the range of the raw data. If this occurs for any year
a warning is issued for the user to check.
}
