% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.data.R
\name{summary.data}
\alias{summary.data}
\title{summary.data}
\usage{
\method{summary}{data}(object, ...)
}
\arguments{
\item{object}{data object}

\item{...}{other parameters passed to \code{summary}}
}
\description{
Summary of data object
}
