% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fitmodel.r
\name{summary.fitmodel}
\alias{summary.fitmodel}
\title{Summarises fitted model.}
\usage{
\method{summary}{fitmodel}(object, ...)
}
\arguments{
\item{object}{fitmodel object}

\item{...}{other arguments to \code{summary}}
}
\description{
summarises an object of class "fitmodel"
}
