% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.regdata.R
\name{summary.regdata}
\alias{summary.regdata}
\title{Summarise regularised data}
\usage{
\method{summary}{regdata}(object, Qcutoff = NULL, ...)
}
\arguments{
\item{object}{regdata object}

\item{Qcutoff}{Qcutoff}

\item{...}{other parameters passed to \code{summary}}
}
\description{
summarises an object of class "regdata"
}
