% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockSigma.R, R/CalcSE.R, R/ConvertFacts.R,
%   R/EstChat.R, R/EstLoad.R, R/EstLoadYS.R, R/ExtractDailyLoads.R,
%   R/Intervals.r
\name{BlockSigma}
\alias{BlockSigma}
\alias{CalcSE}
\alias{ConvertFacts}
\alias{LRE-predict-Internal}
\alias{EstChat}
\alias{EstLoad}
\alias{EstLoadYS}
\alias{ExtractDailyLoads}
\alias{Intervals}
\title{Internal functions for \code{predict.fitmodel} function.}
\usage{
BlockSigma(len, Xdesign, W)

CalcSE(Xdesign, LhatC, alpha1, alpha2, beta1, rho, W, sig2)

ConvertFacts(samp.unit)

EstChat(Qdat, modobj, modobjfix)

EstLoad(
  CQdat,
  Qdat,
  Xdesign,
  W,
  sig2,
  rho,
  alpha1,
  alpha2,
  beta1,
  samp.unit,
  se = TRUE,
  Blocklen = 5000
)

EstLoadYS(
  CQdat,
  Qdat,
  concY,
  modobj,
  samp.unit = samp.unit,
  alpha1,
  alpha2,
  se = TRUE,
  Blocklen = 5000
)

ExtractDailyLoads(Date, Xdesign, Lhat, LhatC, Qhat, error, modparms)

Intervals(est, pvalue)
}
\arguments{
\item{len}{len}

\item{Xdesign}{Design matrix}

\item{W}{Var(betahat)}

\item{LhatC}{Concentration estimate}

\item{alpha1}{measurement error}

\item{alpha2}{error in flow rates}

\item{beta1}{coeficient of log Q in the regression}

\item{rho}{rho}

\item{sig2}{sigma^2}

\item{samp.unit}{sampling unit}

\item{Qdat}{Qdat}

\item{modobj}{modobj}

\item{modobjfix}{modobjfix}

\item{CQdat}{Predicted concentration and flow}

\item{se}{se (Default: TRUE)}

\item{Blocklen}{Block length (Default: 5000)}

\item{concY}{concY}

\item{Date}{Date}

\item{Lhat}{Load estimate}

\item{Qhat}{Flow estimate}

\item{error}{error}

\item{modparms}{modparms}

\item{est}{estimate}

\item{pvalue}{pvalue}
}
\description{
Internal functions used for prediction.
EstChat: estimates concentration.
BlockSigma: estimates the variance-covariance matrix.
CalcSE: calls on C code to calculate the standard error.
ConvertFacts: Internal function that determines the conversion
constant to use for loads and flow.Conversion to Million tonnes:
C = mg/L  Q = m^3/s (K.loads) mg/l x m3/1000 = 1kg  & 1kg = 0.001 t
Conversion to millions of ML (K.flow).
EstLoadYS: Function used in \code{predict.fitmodel} to estimate the
load based on predicted concentrations from predicted hourly flow records.
EstLoad: Function called by \code{EstLoadYS}
ExtractDailyLoads: extracts the daily loads.
Intervals: calculates confidence intervals.
}
