% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periods_of_change.r
\name{periods_of_change}
\alias{periods_of_change}
\title{Calculate and Plot Periods of Change in a non-linear curve}
\usage{
periods_of_change(
  object,
  Term = "trendY",
  data,
  object.vars = c("pQ", "MA2days", "MAweek", "MAmonth", "MA6months", "MA12months",
    "trendY"),
  x_label = NULL,
  spacing = 5,
  nsamps = NULL,
  alpha = 0.05
)
}
\arguments{
\item{object}{gam model produced from the mgcv R package}

\item{Term}{term to calculate the periods of change over, typically a trend term}

\item{data}{data used in the gam model, eg loaddata$CQ}

\item{object.vars}{variables used in the gam fit.}

\item{x_label}{label for the x-axis.  If NULL, it will default to numbers starting at 1.}

\item{spacing}{white space left between the labels. Defaults to 5.}

\item{nsamps}{Typically NULL but can be a vector of the number of samples shown for each point in time.}

\item{alpha}{significance level. Default is 0.05.}
}
\description{
Calculates and plots the episodic trends in a non-linear trend
       curve produced from a gam
}
