% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fitmodel.R
\name{plot.fitmodel}
\alias{plot.fitmodel}
\title{Plot fitmodel objects}
\usage{
\method{plot}{fitmodel}(x, Qreg, ...)
}
\arguments{
\item{x}{data object}

\item{Qreg}{regularised flow}

\item{...}{other arguments to pass plot}
}
\description{
Plot object of class 'fitmodel'
}
